pyproject.toml
[tool.poetry]
name = "create-dc-sw-config"
version = "0.1.0"
description = ""
authors = ["liuyuanchengweb <1046404397@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
pyyaml = "^6.0.2"
jinja2 = "^3.1.4"
pydantic = "^2.8.2"
macaddress = "^2.0.2"
pyinstaller = {version = "^6.10.0", python = ">=3.11,<3.14"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
